[SYNCOPE-1804] Introducing syncope-core-persistence-neo4j (#628)

+++ core/persistence-jpa/src/main/java/org/apache/syncope/core/persistence/jpa/dao/repo/RemediationRepoExtImpl.java
@@ -73,7 +73,7 @@ public class RemediationRepoExtImpl implements RemediationRepoExt {
             query.setParameter("after", after);
         }
 
-        return ((Number) query.getSingleResult()).intValue();
+        return ((Number) query.getSingleResult()).longValue();
     }
 
     @Override
